+++ app/(datasets)/stories/hub.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { StoriesHubContent, useFiltersWithQS } from '@lib';
 import Link from 'next/link';
-
+import Providers from '../providers';
 
 export default function Hub({
   stories: allStories
@@ -12,18 +12,20 @@ export default function Hub({
   const controlVars = useFiltersWithQS();
 
   return (
-    <StoriesHubContent
-      allStories={allStories}
-      onFilterchanges={() => controlVars}
-      storiesString= {{
-        one: 'story',
-        other: 'stories'
-      }}
-      linkProperties={{
-        LinkElement: Link,
-        pathAttributeKeyName: 'href'
-      }}
-      storiesPagePath={'stories'}
-    />
+    <Providers>
+      <StoriesHubContent
+        allStories={allStories}
+        onFilterchanges={() => controlVars}
+        storiesString= {{
+          one: 'story',
+          other: 'stories'
+        }}
+        linkProperties={{
+          LinkElement: Link,
+          pathAttributeKeyName: 'href'
+        }}
+        storiesPagePath={'stories'}
+      />
+    </Providers>
   );
 }
