+++ app/(datasets)/stories/page.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { Suspense } from 'react';
 import { getStoriesMetadata } from 'app/content/utils/mdx';
 import Hub from './hub';
 
@@ -9,10 +9,13 @@ export default function Page() {
   }));
 
   return (
-    <section>
-      <div className='grid-container'>
-        <Hub stories={stories} />
-      </div>
-    </section>
+    <div className='grid-container'>
+      <section>
+        <h1 className='font-ui-lg'>Stories</h1>
+        <Suspense fallback={<>Loading...</>}>
+          <Hub stories={stories} />
+        </Suspense>
+      </section>
+    </div>
   );
 }
